% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_stoichiometry}
\alias{get_stoichiometry}
\title{Relative stoichiometry}
\usage{
get_stoichiometry(dep, ibaq, contrast, bait, level = 1)
}
\arguments{
\item{dep}{SummarizedExperiment,
Proteomics dataset in which differential enriched proteins
are annotated by \code{\link[DEP]{add_rejections}}.}

\item{ibaq}{Data.frame,
iBAQ table generated by \code{\link{merge_ibaq}}.}

\item{contrast}{Character(1),
The specific contrast for which to calculate the stoichiometries.}

\item{bait}{Character(1),
The name of the protein to which all other proteins will be scaled.}

\item{level}{Numerical(1),
The level to which the bait will be scaled.}
}
\value{
A data.frame with relative stoichiometry values.
}
\description{
\code{get_stoichiometry} calculates the relative stoichiometries of
all differentially enriched proteins.
}
\examples{
# load data
proteins <- GFPip
proteins <- data_unique <- make_unique(proteins, 'Gene.names', 'Protein.IDs')
exp_design <- GFPip_ExpDesign
peptides <- GFPip_pep

# Test for differential enriched proteins
se <- import_MaxQuant(proteins, exp_design,
    filter = c("Reverse", "Contaminant"))
processed <- process(se, fun = "MinProb")
dep <- analyze_dep(processed, 'WT', 'control', lfc = 4.5)

# Merge iBAQ intensities of proteins that have shared peptides
ibaq <- merge_ibaq(data_unique, peptides)

# Calculate relative stoichiometry versus 'Suz12' in the 'GFP_vs_WT' contrast
stoi <- get_stoichiometry(dep, ibaq, contrast = 'GFP_vs_WT', bait = 'Suz12')

}
