% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_ibaq}
\alias{plot_ibaq}
\title{Plot iBAQ values versus log fold change}
\usage{
plot_ibaq(dep, contrast, labelsize = 3)
}
\arguments{
\item{dep}{SummarizedExperiment object,
Proteomics dataset on which differential enriched proteins
are annotated by \code{\link[DEP]{add_rejections}}.}

\item{contrast}{Character(1),
The specific contrast to plot.}

\item{labelsize}{Integer(1),
Sets the size of name labels.}
}
\value{
A scatter plot
}
\description{
\code{plot_ibaq} plots a scatter plot of the
iBAQ intensities versus the LFQ fold changes.
}
\examples{
# load data
proteins <- GFPip
exp_design <- GFPip_ExpDesign

# Test for differential enriched proteins
se <- import_MaxQuant(proteins, exp_design,
    filter = c("Reverse", "Contaminant"))
processed <- process(se, fun = "MinProb")
dep <- analyze_dep(processed, 'WT', 'control', lfc = 4.5)

# Plot iBAQ vs LFQ plot
plot_ibaq(dep, 'GFP_vs_WT', labelsize = 3)

}
