% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_stoichiometry}
\alias{plot_stoichiometry}
\title{Plot relative stoichiometries}
\usage{
plot_stoichiometry(protein_stoichiometry, thr = 0.01, max_y = NULL)
}
\arguments{
\item{protein_stoichiometry}{Data.frame,
Stoichiometry table generated by \code{\link{get_stoichiometry}}.}

\item{thr}{Numerical(1),
The stoichiometry threshold above which proteins will be plotted.}

\item{max_y}{Numerical(1),
The maximum scale of the y-axis.}
}
\value{
A barplot (generated by \code{\link[ggplot2]{ggplot}}).
}
\description{
\code{plot_stoichiometry} plots a barplot of the
relative stoichiometries.
}
\examples{
# load data
proteins <- GFPip
proteins <- data_unique <- make_unique(proteins, 'Gene.names', 'Protein.IDs')
exp_design <- GFPip_ExpDesign
peptides <- GFPip_pep

# Test for differential enriched proteins
se <- import_MaxQuant(proteins, exp_design,
    filter = c("Reverse", "Contaminant"))
processed <- process(se, fun = "MinProb")
dep <- analyze_dep(processed, 'control', 'WT', lfc = 4.5)

# Merge iBAQ intensities of proteins that have shared peptides
ibaq <- merge_ibaq(data_unique, peptides)

# Calculate relative stoichiometry versus 'Suz12' in the 'GFP_vs_WT' contrast
stoi <- get_stoichiometry(dep, ibaq, contrast = 'GFP_vs_WT', bait = 'Suz12')

# Plot stoichiometry
plot_stoichiometry(stoi)

}
